requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.4
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
psycopg2-binary==2.9.9
redis==5.0.0
gunicorn==21.2.0
dj-database-url==2.1.0
black==25.1.0
